Refactor equation vector
